+++ pkg/eval/compile_value_test.go
@@ -193,7 +193,7 @@ func TestClosure(t *testing.T) {
 		// Closure captures new local variables every time
 		That(`fn f []{ x=0; put []{x=(+ $x 1)} []{put $x} }
 		  {inc1,put1}=(f); $put1; $inc1; $put1
-		  {inc2,put2}=(f); $put2; $inc2; $put2`).Puts("0", 1.0, "0", 1.0),
+		  {inc2,put2}=(f); $put2; $inc2; $put2`).Puts("0", int64(1), "0", int64(1)),
 
 		// Rest argument.
 		That("[x @xs]{ put $x $xs } a b c").Puts("a", vals.MakeList("b", "c")),
